Add files via upload
Displays a list of the order and type of client the clerk receives, as well as when the clients are helped. There are two types of clients - in person clients and over-the-phone clients. In-person clients must wait in the back of a line to be helped. Over-the-phone clients are helped immediately. New clients are received and current clients are helped based on semi-random time intervals. The program runs for 10 simulated seconds before the ordered list is displayed to the user.